Move some signal-handling code around
