+++ src/twisted/internet/posixbase.py
@@ -7,14 +7,11 @@ Posix reactor base class
 """
 
 
-import contextlib
-import errno
-import os
 import socket
 import sys
 from typing import Sequence
 
-from zope.interface import Attribute, Interface, classImplements, implementer
+from zope.interface import classImplements, implementer
 
 from twisted.internet import error, tcp, udp
 from twisted.internet.base import ReactorBase, _SignalReactorMixin
@@ -31,7 +28,7 @@ from twisted.internet.interfaces import (
     IReactorUNIXDatagram,
 )
 from twisted.internet.main import CONNECTION_DONE, CONNECTION_LOST
-from twisted.python import failure, log, util
+from twisted.python import failure, log
 from twisted.python.runtime import platform, platformType
 
 # Exceptions that doSelect might return frequently
@@ -57,7 +54,8 @@ unixEnabled = platformType == "posix"
 
 processEnabled = False
 if unixEnabled:
-    from twisted.internet import _signals, fdesc, process, unix
+    from twisted.internet import process, unix
+    from ._signals import _SIGCHLDWaker, _Waker
 
     processEnabled = True
 
@@ -69,200 +67,6 @@ if platform.isWindows():
         processEnabled = True
     except ImportError:
         win32process = None
-
-
-class _IWaker(Interface):
-    """
-    Interface to wake up the event loop based on the self-pipe trick.
-
-    The U{I{self-pipe trick}<http://cr.yp.to/docs/selfpipe.html>}, used to wake
-    up the main loop from another thread or a signal handler.
-    This is why we have wakeUp together with doRead
-
-    This is used by threads or signals to wake up the event loop.
-    """
-
-    disconnected = Attribute("")
-
-    def wakeUp():
-        """
-        Called when the event should be wake up.
-        """
-
-    def doRead():
-        """
-        Read some data from my connection and discard it.
-        """
-
-    def connectionLost(reason: failure.Failure):
-        """
-        Called when connection was closed and the pipes.
-        """
-
-
-@implementer(_IWaker)
-class _SocketWaker(log.Logger):
-    """
-    The I{self-pipe trick<http://cr.yp.to/docs/selfpipe.html>}, implemented
-    using a pair of sockets rather than pipes (due to the lack of support in
-    select() on Windows for pipes), used to wake up the main loop from
-    another thread.
-    """
-
-    disconnected = 0
-
-    def __init__(self, reactor):
-        """Initialize."""
-        self.reactor = reactor
-        # Following select_trigger (from asyncore)'s example;
-        client = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        client.setsockopt(socket.IPPROTO_TCP, socket.TCP_NODELAY, 1)
-        with contextlib.closing(
-            socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        ) as server:
-            server.bind(("127.0.0.1", 0))
-            server.listen(1)
-            client.connect(server.getsockname())
-            reader, clientaddr = server.accept()
-        client.setblocking(0)
-        reader.setblocking(0)
-        self.r = reader
-        self.w = client
-        self.fileno = self.r.fileno
-
-    def wakeUp(self):
-        """Send a byte to my connection."""
-        try:
-            util.untilConcludes(self.w.send, b"x")
-        except OSError as e:
-            if e.args[0] != errno.WSAEWOULDBLOCK:
-                raise
-
-    def doRead(self):
-        """
-        Read some data from my connection.
-        """
-        try:
-            self.r.recv(8192)
-        except OSError:
-            pass
-
-    def connectionLost(self, reason):
-        self.r.close()
-        self.w.close()
-
-
-class _FDWaker(log.Logger):
-    """
-    The I{self-pipe trick<http://cr.yp.to/docs/selfpipe.html>}, used to wake
-    up the main loop from another thread or a signal handler.
-
-    L{_FDWaker} is a base class for waker implementations based on
-    writing to a pipe being monitored by the reactor.
-
-    @ivar o: The file descriptor for the end of the pipe which can be
-        written to wake up a reactor monitoring this waker.
-
-    @ivar i: The file descriptor which should be monitored in order to
-        be awoken by this waker.
-    """
-
-    disconnected = 0
-
-    i = None
-    o = None
-
-    def __init__(self, reactor):
-        """Initialize."""
-        self.reactor = reactor
-        self.i, self.o = os.pipe()
-        fdesc.setNonBlocking(self.i)
-        fdesc._setCloseOnExec(self.i)
-        fdesc.setNonBlocking(self.o)
-        fdesc._setCloseOnExec(self.o)
-        self.fileno = lambda: self.i
-
-    def doRead(self):
-        """
-        Read some bytes from the pipe and discard them.
-        """
-        fdesc.readFromFD(self.fileno(), lambda data: None)
-
-    def connectionLost(self, reason):
-        """Close both ends of my pipe."""
-        if not hasattr(self, "o"):
-            return
-        for fd in self.i, self.o:
-            try:
-                os.close(fd)
-            except OSError:
-                pass
-        del self.i, self.o
-
-
-@implementer(_IWaker)
-class _UnixWaker(_FDWaker):
-    """
-    This class provides a simple interface to wake up the event loop.
-
-    This is used by threads or signals to wake up the event loop.
-    """
-
-    def wakeUp(self):
-        """Write one byte to the pipe, and flush it."""
-        # We don't use fdesc.writeToFD since we need to distinguish
-        # between EINTR (try again) and EAGAIN (do nothing).
-        if self.o is not None:
-            try:
-                util.untilConcludes(os.write, self.o, b"x")
-            except OSError as e:
-                # XXX There is no unit test for raising the exception
-                # for other errnos. See #4285.
-                if e.errno != errno.EAGAIN:
-                    raise
-
-
-if platformType == "posix":
-    _Waker = _UnixWaker
-else:
-    # Primarily Windows and Jython.
-    _Waker = _SocketWaker  # type: ignore[misc,assignment]
-
-
-class _SIGCHLDWaker(_FDWaker):
-    """
-    L{_SIGCHLDWaker} can wake up a reactor whenever C{SIGCHLD} is
-    received.
-
-    @see: L{twisted.internet._signals}
-    """
-
-    def __init__(self, reactor):
-        _FDWaker.__init__(self, reactor)
-
-    def install(self):
-        """
-        Install the handler necessary to make this waker active.
-        """
-        _signals.installHandler(self.o)
-
-    def uninstall(self):
-        """
-        Remove the handler which makes this waker active.
-        """
-        _signals.installHandler(-1)
-
-    def doRead(self):
-        """
-        Having woken up the reactor in response to receipt of
-        C{SIGCHLD}, reap the process which exited.
-
-        This is called whenever the reactor notices the waker pipe is
-        writeable, which happens soon after any call to the C{wakeUp}
-        method.
-        """
-        _FDWaker.doRead(self)
-        process.reapAllProcesses()
 
 
 class _DisconnectSelectableMixin:
